Feat: Adding PlaceCard and Shimmer components
Summary: Touchable was added but it should come from universal-components.
PlaceCard comes with its own Placeholder component, which relies on Shimmer component; it is a css-solution for web and a wrapper of `react-native-shimmer-placeholder` for RN.
Changed jest's preset to `jest-expo` as `react-native` was not enough.
Edited defaultSource and source props for Image component in react-native-web flow-typed types: their entries are now read-only.