Chore: update @kiwicom/graphql-utils to v0.3.0
Used OptimisticDataloader from @kiwicom/graphql-utils
Original data-loader kills load batch if one of this response values is
Error instance. This optimistic data-loader keeps Error value as a valid
response. This means that it's possible to return as many responses
as possible and convert Errors to the error messages in the response
if necessary.

 @see https://github.com/facebook/dataloader/issues/41