Feat: Added visual regression testing
Summary: Like on `orbit-components`, we use [Loki](https://loki.js.org) to perform visual testing.
Instructions to run the tests were added to CONTRIBUTING, as well as how to let Loki skip stories which are too dependent on context for them to always look the same. A good example is the Playground for the DatePicker component; a new date is generated at each render.
This PR also adds a .loki folder at the root, containing screenshots of all the (non-skipped) stories we wrote.