Feat: Added serverless deployment config
Summary: Currently, the deployment process is manual. Adding it to CircleCI will be done in a separate PR.
I removed the function build and deploy on Netlify.

Assuming you have AWS credentials with aws-cli configured and that you are in apps/graphql, you would just need to do:
- `yarn webpack`: this builds the server into a Node-compatible JavaScript
- `serverless deploy function --function margarita-graphql`: this will update the Lambda function

We could probably add a test to make sure the deployment was successful, by sending a POST request to the endpoint for example. If it did not work, it should somehow rollback (though I am not sure how that works precisely)