Fix: (Partial) Hydrate props from query params
Summary: When going to /results page from a built nextjs application, getInitialProps and the likes are ignored. That meant that refreshing /results would return Error because the query params were not taken into account.

I used the solution from this comment https://github.com/zeit/next.js/issues/4804#issuecomment-451558569, and it works pretty well.

Caveat is that we show for a brief moment a sort of broken state. That should be improved.