Chore: Use Text component wrapper in @kiwicom/margarita-components
Summary: This is a follow up PR to https://github.com/kiwicom/margarita/pull/248; we should use this Text wrapper component to ensure that `fontWeight` bold and `fontStyle` italic get applied correctly on iOS and Android through Expo.

Note that we should use `weight='bold'` and `italic` props instead of passing them through `style` prop (unless we need the text to be bold only on `web` for example).