Fix: Flow types work in host project
Summary: Flow types were not picked up when exported as an NPM package; changing "main" and "react-native" fields in package.json to point to the right index.js files does the trick. I suspect that the previous setup was masking flow types because we were using module.exports without declaring a type for the module.