Chore: Upgrading @kiwicom/universal-components v0.0.13
Summary: Now that Flow types are correctly being picked up, there were quite a few Flow errors to fix.
Some are related to using react-native-web flow-typed in this repo and they conflict with `@kiwicom/universal-components` in `node_modules`. These will have to be fixed there and they are ignored in this repo in .flowconfig file.
There was also an issue with how Jest resolves `@kiwicom/universal-components`: it only looks at the `main` field in `package.json`, and not at `react-native` field if preset is `jest-expo` or `react-native` in the config file`.jest.json`. That means that Jest resolved the web version of the components instead of the RN version.