Feat: Display warning if no locations
Summary: The current behaviour is to display the message only if both travelFrom and travelTo are null.
If you want to reproduce it you can replace in `SearchContext.js`
```
const defaultState = {
  tripType: 'return',
  travelFrom: null, // defaultPlaces.origin,
  travelTo: null, // defaultPlaces.departure,
...
```

This PR adds a new Alert component (together with a context to use it) in @kiwicom/margarita-components. It displays a Modal UI element on web and uses react-native's Alert on mobile.

Closes https://github.com/kiwicom/margarita/issues/276