Chore: No automatic push to GitHub
Summary: By pushing to GitHub during NPM publishing, the `package.json`'s "main" entrypoint point to `lib/web/index.js` and not `src/index.js`, which is not what we want. The adjusted release-it config and release-it command prevents commiting, pushing and tagging.
That also means we will need to manually commit the updated version to the repo. To be improved further.