Feat: Added code to Carrier type
Summary: 
Renamed `Transporter` type to `Carrier`

So far only name was used for Transporter but in a wrong way in multiple cases, because instead of actual carrier name the code was mapped to this value.
Because both values are necessary for result detail, `code` value was added and mapping with implementation was updated.