Docs: Readme update
Current readme file is outdated and full of TODOs. But in fact we already have most of these sections covered in documentation.

To prevent duplicate data (and need to maintain it in different places) readme was currently reduced to `demo` links, `documentation` link, few important `development` related links and `license`.

We can extend it later with some better project explanation. Once we obtain this content/copy for landing page.