Docs: Cleaned 'What is this all about?' section (#728)
For now this section was just copy of repo readme. Most of the info is already part of the other sections, so it is no longer necessary to duplicate it there.

As with recently updated repo readme, this section was just stripped to few relevant informations and we can extend it later when we put together some catchy copy for landing page.